requirements.txt
Django==5.1.1
pybit==5.8.0
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
